Enabled "method+instance" registration of callbacks in policies and blocks as it is for dataflows too. Also fixed 3 warnings

+++ flow/include/flow/ThreadPool.h
@@ -49,7 +49,7 @@ namespace flow{
         ~ThreadPool();
       
         /// Get size of task queue
-        unsigned queueSize() {
+        size_t queueSize() {
             return tasks_.size();
         }
